Move definition loading code into its own module
We're going to be making this more complex soon so it will be helpful to
separate it out.
